refactor(webui): decouple slash command highlighting
