+++ webui/src/components/SlashCommandText.tsx
@@ -1,43 +1,22 @@
 import {
-  CliAppMentionText,
   INLINE_TOKEN_HIGHLIGHT_COLOR,
   InlineTokenHighlight,
-} from "@/components/CliAppMentionText";
-import { matchingSlashCommand } from "@/lib/slash-command";
-import type { CliAppInfo, McpPresetInfo, SlashCommand } from "@/lib/types";
+} from "@/components/InlineTokenHighlight";
 
 interface SlashCommandTextProps {
-  text: string;
-  slashCommands: SlashCommand[];
-  cliApps: CliAppInfo[];
-  mcpPresets: McpPresetInfo[];
+  command: string;
 }
 
 export function SlashCommandText({
-  text,
-  slashCommands,
-  cliApps,
-  mcpPresets,
+  command,
 }: SlashCommandTextProps) {
-  const command = matchingSlashCommand(text, slashCommands);
-  if (!command) {
-    return <CliAppMentionText text={text} cliApps={cliApps} mcpPresets={mcpPresets} />;
-  }
-
   return (
-    <>
-      <InlineTokenHighlight
-        testId="message-slash-command"
-        color={INLINE_TOKEN_HIGHLIGHT_COLOR}
-        className="font-medium"
-      >
-        {command.command}
-      </InlineTokenHighlight>
-      <CliAppMentionText
-        text={text.slice(command.command.length)}
-        cliApps={cliApps}
-        mcpPresets={mcpPresets}
-      />
-    </>
+    <InlineTokenHighlight
+      testId="message-slash-command"
+      color={INLINE_TOKEN_HIGHLIGHT_COLOR}
+      className="font-medium"
+    >
+      {command}
+    </InlineTokenHighlight>
   );
 }
